+++ fixtures/_test_hot_reload/web/main.dart
@@ -7,11 +7,17 @@ import 'dart:js_interop';
 
 import 'package:_test_hot_reload/library1.dart';
 
+@JS('console.log')
+external void log(String _);
+
 @JS('document.body.innerHTML')
 external set innerHtml(String html);
 
 void evaluate() {
   innerHtml = 'Program is running!\n $reloadValue}\n';
+
+  // Wait for this print statement so that we know evaluate is done executing.
+  log('evaluate executed');
 }
 
 void main() {